fix template reload bug
