+++ pkg/microservice/aslan/core/common/types/service_variable_test.go
